backend_services/requirements.txt
djangorestframework
markdown
django-filter
python-decouple
django-simple-history
django-cors-headers==4.7.0
django-filter==25.1
drf-spectacular
